Avoid use of reflectmethod attribute in Go1.22
In a future version of Go, a compiler assist is added to special case
uses of reflect.Value.MethodByName() where the string argument is a
constant.

https://go-review.googlesource.com/c/go/+/522436/5

This avoids marking the calling function as <ReflectMethod>, which
can result in binary bloat simply by importing the storage package,
as the linker is unable to prove which methods are reachable in the
final binary. This can result in significant binary size reduction and
related performance benefits.

Refer:

https://docs.google.com/document/d/1KtJSpRvoHddya-P329ZgLwkv-NkyCXCRMX3FOAeDoBk/edit#heading=h.ltobgcan4ct3

for further context.